ReclaimMannequinCommandBaseHandler
