+++ src/Octoshift/Commands/ReclaimMannequinCommandBase.cs
@@ -1,22 +1,13 @@
 using System;
 using System.CommandLine;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using Octoshift;
 using OctoshiftCLI.Contracts;
-using OctoshiftCLI.Extensions;
 
 namespace OctoshiftCLI.Commands;
 
 public class ReclaimMannequinCommandBase : Command
 {
-    private readonly OctoLogger _log;
-    private readonly ITargetGithubApiFactory _githubApiFactory;
-    private ReclaimService _reclaimService;
-
-    internal Func<string, bool> FileExists = path => File.Exists(path);
-    internal Func<string, string[]> GetFileContent = path => File.ReadLines(path).ToArray();
+    protected ReclaimMannequinCommandBaseHandler BaseHandler { get; init; }
 
     public ReclaimMannequinCommandBase(OctoLogger log, ITargetGithubApiFactory githubApiFactory, ReclaimService reclaimService = null) : base(
         name: "reclaim-mannequin",
@@ -29,9 +20,7 @@ public class ReclaimMannequinCommandBase : Command
                      Environment.NewLine +
                      "If both options are specified The CSV file takes precedence and other options will be ignored")
     {
-        _log = log;
-        _githubApiFactory = githubApiFactory;
-        _reclaimService = reclaimService;
+        BaseHandler = new ReclaimMannequinCommandBaseHandler(log, githubApiFactory, reclaimService);
     }
 
     protected virtual Option<string> GithubOrg { get; } = new("--github-org")
@@ -90,61 +79,6 @@ public class ReclaimMannequinCommandBase : Command
         AddOption(Force);
         AddOption(GithubPat);
         AddOption(Verbose);
-    }
-
-    public async Task Handle(ReclaimMannequinCommandArgs args)
-    {
-        if (args is null)
-        {
-            throw new ArgumentNullException(nameof(args));
-        }
-
-        _log.Verbose = args.Verbose;
-
-        if (string.IsNullOrEmpty(args.Csv) && (string.IsNullOrEmpty(args.MannequinUser) || string.IsNullOrEmpty(args.TargetUser)))
-        {
-            throw new OctoshiftCliException($"Either --{Csv.ArgumentHelpName} or --{MannequinUsername.ArgumentHelpName} and --{TargetUsername.ArgumentHelpName} must be specified");
-        }
-
-        var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: args.GithubPat);
-        _reclaimService ??= new ReclaimService(githubApi, _log);
-
-        if (!string.IsNullOrEmpty(args.Csv))
-        {
-            _log.LogInformation("Reclaiming Mannequins with CSV...");
-
-            _log.LogInformation($"{GithubOrg.GetLogFriendlyName()}: {args.GithubOrg}");
-            _log.LogInformation($"FILE: {args.Csv}");
-            if (args.Force)
-            {
-                _log.LogInformation("MAPPING RECLAIMED");
-            }
-
-            if (!FileExists(args.Csv))
-            {
-                throw new OctoshiftCliException($"File {args.Csv} does not exist.");
-            }
-
-            await _reclaimService.ReclaimMannequins(GetFileContent(args.Csv), args.GithubOrg, args.Force);
-        }
-        else
-        {
-            _log.LogInformation("Reclaiming Mannequin...");
-
-            _log.LogInformation($"{GithubOrg.GetLogFriendlyName()}: {args.GithubOrg}");
-            _log.LogInformation($"MANNEQUIN: {args.MannequinUser}");
-            if (args.MannequinId != null)
-            {
-                _log.LogInformation($"{MannequinId.GetLogFriendlyName()}: {args.MannequinId}");
-            }
-            _log.LogInformation($"RECLAIMING USER: {args.TargetUser}");
-            if (args.GithubPat is not null)
-            {
-                _log.LogInformation($"{GithubPat.GetLogFriendlyName()}: ***");
-            }
-
-            await _reclaimService.ReclaimMannequin(args.MannequinUser, args.MannequinId, args.TargetUser, args.GithubOrg, args.Force);
-        }
     }
 }
