+++ src/OctoshiftCLI.Tests/Octoshift/Commands/ReclaimMannequinCommandBaseTests.cs
@@ -15,7 +15,7 @@ public class ReclaimMannequinCommandBaseTests
     private readonly Mock<ITargetGithubApiFactory> _mockTargetGithubApiFactory = new();
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
     private readonly Mock<ReclaimService> _mockReclaimService = TestHelpers.CreateMock<ReclaimService>();
-    private readonly ReclaimMannequinCommandBase _command;
+    private readonly ReclaimMannequinCommandBaseHandler _command;
 
     private const string GITHUB_ORG = "FooOrg";
     private const string MANNEQUIN_USER = "mona";
@@ -23,7 +23,7 @@ public class ReclaimMannequinCommandBaseTests
 
     public ReclaimMannequinCommandBaseTests()
     {
-        _command = new ReclaimMannequinCommandBase(_mockOctoLogger.Object, _mockTargetGithubApiFactory.Object, _mockReclaimService.Object)
+        _command = new ReclaimMannequinCommandBaseHandler(_mockOctoLogger.Object, _mockTargetGithubApiFactory.Object, _mockReclaimService.Object)
         {
             FileExists = _ => true,
             GetFileContent = _ => Array.Empty<string>()
